Fix CI problem
On GitHub Actions, we cannot access to GitHub by SSH, as we have no
authorized access key.
On new zsh process, when zsh reads .zshrc and executes git clone zinit
project, git access to GitHub with SSH because of the configuration on
~/.gitconfig.

It would be better to find a way not to use default ~/.gitconfig while
on deploying process.